+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Next
+- Fix duplication of points on child borders when node get divided
+
 ## 3.3.0
 - Insertion returns a boolean (true if inserted successfully, false otherwise)